feat & remove : followup_question_manager.py 삭제 후, answer_service.py에 로직 이관. 그리고 answer_service.py 스켈레톤 코드 작성.
